Simplify the main loop of KeyRangeIntersectionIterator
There is a bit more code in total, but IMHO it is more obvious now.
The main simplification stems from removing the responsibility
of the initial advancement of an iterator from the main loop, and
moving that to a separate step before the loop.

The first step is now very explicit about advancing the non-STATIC
key first (if it exists), before the STATIC ones.

Now the main loop is responsible only for matching the lagging
iterators. Hence, we can have a nice direct stopping condition
for the main  loop - it stops when all iterators are at the same key.
The decision whether the iterator should be advanced or not is
made only based on how it compares to the highest key,
not whether it is STATIC or WIDE. This could be done because a STATIC
key is equal to all WIDE keys in the same partition.
We could also remove one variable remembering which iterator
was advanced last.